Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,20 +1,20 @@
 #include <stdio.h>
 int main() {
-    float double a,b;
+    float a,b;
     char c;
     scanf("%d%d%c", &a,&b,&c);
     switch (c){
         case '+':
-        printf("%lf",a+b);
+        printf("%f",a+b);
         break;
         case '-':
-        printf("%lf",a-b);
+        printf("%f",a-b);
         break;
         case '*':
-        printf("%lf",a*b);
+        printf("%f",a*b);
         break;
         case '/':
-        printf("%lf",a/b);
+        printf("%f",a/b);
         break;
         default:
         printf("Invalid Input");
